api/requirements.txt
fastapi
uvicorn
asyncpg
jinja2
python-multipart
psycopg2-binary